server/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
openenv-core==0.2.3
pydantic==2.12.5
pandas==3.0.2
numpy==2.4.4
rapidfuzz==3.14.3
openai==2.30.0
requests==2.33.1
python-dotenv==1.2.2
python-multipart==0.0.22
requests
